Check that image has been successfully flashed.
Instead of using memcpy to copy the source data to a RAM buffer ready
for flashing, use DMA instead.  This will be slightly faster but also
allows us to calculate the CRC of the copied data at the same time.
The CRC is then compared against the CRC of the flashed data to ensure
they are both the same.